# Generalize RungeKutta3 and QuasiAdamsBashforth2 for any model