--- a/src/Models/IncompressibleModels/IncompressibleModels.jl
+++ b/src/Models/IncompressibleModels/IncompressibleModels.jl
@@ -5,7 +5,7 @@ using KernelAbstractions.Extras.LoopInfo: @unroll
 
 using Oceananigans.Utils: launch!
 
-import ..Models: fields
+import Oceananigans: fields
 
 #####
 ##### IncompressibleModel definition