--- a/src/Models/ShallowWaterModels/ShallowWaterModels.jl
+++ b/src/Models/ShallowWaterModels/ShallowWaterModels.jl
@@ -30,8 +30,4 @@ include("solution_and_tracer_tendencies.jl")
 include("calculate_shallow_water_tendencies.jl")
 include("update_shallow_water_state.jl")
 
-# These files can be removed when rk3_substep! and store_tendencies! are generalized:
-#include("rk3_substep_shallow_water_model.jl")
-#include("store_shallow_water_tendencies.jl")
-
 end # module