This generalized the runge_kutta3.jl script to work for any model.  Also, it combines the time stepping for the fields, say velocities, and the tracers into one loop.  Thanks to @glwagner and @ali-ramadhan !

I do get the following warnings that I would like to eliminate. Any suggestions how i can do that?

```
WARNING: ignoring conflicting import of Models.fields into IncompressibleModels
WARNING: using Models.fields in module Oceananigans conflicts with an existing identifier.

```