@francispoulin 2020-11-25
Given some checks were not successful I now see that there are some problems with this PR.  I glanced at the errors but I don't prettend to understand what's going on.