@ali-ramadhan 2020-11-25
Ah I think the method conflict warnings were because `function fields end` was being defined in `src/Oceananigans.jl` and in `src/Models/Models.jl`. I think you meant to only define it in `src/Oceananigans.jl` since it's needed in `src/TimeSteppers.jl` (which is defined before `src/Models/Models.jl`).

By the time I figured this out I already modified quite a bit so I just pushed my changes, but warnings should be gone now!