@ali-ramadhan 2020-11-25
I think the tests almost pass! It's just a few tests in `test_abstract_operations.jl` that don't: https://buildkite.com/clima/oceananigans/builds/670#3ec51acd-7496-449e-afb6-ace178715cf3/14-394

Maybe they're just missing a `using Oceananigans: fields` or maybe `fields` needs to be exported from the `Oceananigans` and/or `Models` modules?