@francispoulin 2020-11-25
Thanks @ali-ramadhan for finding the problem and fixing it. 

I will pull the updated code and try the test that you pointed out.  Not sure I will be able to figure it out or not though but I will try.