@francispoulin 2020-11-26
> Looks great! Less code, more features too.
> 
> I think there are some conflicts with the master branch so you might have to `git merge master` into this branch and resolve the conflicts locally and push the changes before the PR can be merged.
> 
> Sometimes the conflicts are pretty minor in which case you could resolve them directly on GitHub. But in this case GitHub says they're too complex.

Sounds good.  Luckily those are files that I created.  I was planning on fixing them up tomorrow anyhow, now theres just added incentive.