@francispoulin 2020-11-26
For the new branch I get the following results from benchmark_incompressible_model

                                        Incompressible model benchmarks
┌───────────────┬─────────────┬─────┬────────────┬────────────┬────────────┬────────────┬────────────┬────────┐
│ Architectures │ Float_types │  Ns │        min │     median │       mean │        max │     memory │ allocs │
├───────────────┼─────────────┼─────┼────────────┼────────────┼────────────┼────────────┼────────────┼────────┤
│           CPU │     Float32 │  32 │   3.770 ms │   3.925 ms │   3.975 ms │   4.535 ms │ 247.69 KiB │   1916 │
│           CPU │     Float32 │  64 │  24.751 ms │  24.945 ms │  25.124 ms │  26.909 ms │ 247.69 KiB │   1916 │
│           CPU │     Float32 │ 128 │ 218.012 ms │ 218.721 ms │ 219.037 ms │ 220.987 ms │ 247.69 KiB │   1916 │
│           CPU │     Float64 │  32 │   4.253 ms │   4.437 ms │   4.509 ms │   5.229 ms │ 299.80 KiB │   1916 │
│           CPU │     Float64 │  64 │  29.137 ms │  29.446 ms │  29.689 ms │  31.794 ms │ 299.80 KiB │   1916 │
│           CPU │     Float64 │ 128 │ 257.251 ms │ 258.619 ms │ 259.852 ms │ 270.451 ms │ 299.80 KiB │   1916 │
│           GPU │     Float32 │  32 │   2.489 ms │   2.591 ms │   2.755 ms │   3.150 ms │ 814.41 KiB │  11740 │
│           GPU │     Float32 │  64 │  10.374 ms │  13.950 ms │  13.590 ms │  14.010 ms │ 814.38 KiB │  11746 │
│           GPU │     Float32 │ 128 │  88.020 ms │ 125.190 ms │ 122.408 ms │ 133.906 ms │ 814.38 KiB │  11746 │
│           GPU │     Float64 │  32 │   5.323 ms │   5.438 ms │   5.431 ms │   5.573 ms │ 892.33 KiB │  11574 │
│           GPU │     Float64 │  64 │  34.741 ms │  43.748 ms │  42.586 ms │  44.978 ms │ 892.30 KiB │  11580 │
│           GPU │     Float64 │ 128 │ 279.110 ms │ 333.392 ms │ 328.209 ms │ 335.085 ms │ 892.30 KiB │  11580 │
└───────────────┴─────────────┴─────┴────────────┴────────────┴────────────┴────────────┴────────────┴────────┘
[2020/11/26 16:03:50.829] INFO  Writing Incompressible_model_benchmarks.html...
      Incompressible model CPU -> GPU speedup
┌─────────────┬─────┬──────────┬─────────┬─────────┐
│ Float_types │  Ns │  speedup │  memory │  allocs │
├─────────────┼─────┼──────────┼─────────┼─────────┤
│     Float32 │  32 │  1.51499 │ 3.28804 │ 6.12735 │
│     Float32 │  64 │  1.78816 │ 3.28791 │ 6.13048 │
│     Float32 │ 128 │   1.7471 │ 3.28791 │ 6.13048 │
│     Float64 │  32 │  0.81598 │ 2.97644 │ 6.04071 │
│     Float64 │  64 │ 0.673082 │ 2.97634 │ 6.04384 │
│     Float64 │ 128 │ 0.775721 │ 2.97634 │ 6.04384 │
└─────────────┴─────┴──────────┴─────────┴─────────┘