@francispoulin 2020-11-26
New branch **fjp/generalize-runge-kutta-3**:

```
Coming soon to a PR near you!
```

Old branch **master**:

```
                             Multithreading benchmarks
┌──────┬─────────┬──────────┬──────────┬──────────┬──────────┬────────────┬──────────┐
│ size │ threads │      min │   median │     mean │      max │     memory │   allocs │
├──────┼─────────┼──────────┼──────────┼──────────┼──────────┼────────────┼──────────┤
│  512 │       1 │ 24.654 s │ 24.654 s │ 24.654 s │ 24.654 s │ 294.28 KiB │     1930 │
│  512 │       2 │ 21.380 s │ 21.380 s │ 21.380 s │ 21.380 s │ 172.25 MiB │ 11250274 │
│  512 │       4 │  9.585 s │  9.585 s │  9.585 s │  9.585 s │  63.01 MiB │  4093201 │
│  512 │       8 │  5.417 s │  5.417 s │  5.417 s │  5.417 s │  34.91 MiB │  2242285 │
│  512 │      16 │  3.989 s │  3.991 s │  3.991 s │  3.993 s │ 123.02 MiB │  2196707 │
│  512 │      32 │  3.655 s │  3.676 s │  3.676 s │  3.698 s │  11.86 MiB │   663272 │
│  512 │      36 │  3.783 s │  3.794 s │  3.794 s │  3.804 s │ 115.48 MiB │  1646037 │
└──────┴─────────┴──────────┴──────────┴──────────┴──────────┴────────────┴──────────┘
[2020/11/26 16:58:13.592] INFO  Writing Multithreading_benchmarks.html...
             Multithreading speedup
┌──────┬─────────┬─────────┬─────────┬─────────┐
│ size │ threads │ speedup │  memory │  allocs │
├──────┼─────────┼─────────┼─────────┼─────────┤
│  512 │       1 │     1.0 │     1.0 │     1.0 │
│  512 │       2 │ 1.15317 │ 599.384 │ 5829.16 │
│  512 │       4 │ 2.57214 │ 219.267 │ 2120.83 │
│  512 │       8 │ 4.55103 │ 121.466 │ 1161.81 │
│  512 │      16 │ 6.17726 │ 428.071 │ 1138.19 │
│  512 │      32 │ 6.70601 │ 41.2683 │ 343.664 │
│  512 │      36 │ 6.49899 │ 401.838 │ 852.869 │
└──────┴─────────┴─────────┴─────────┴─────────┘

```