@ali-ramadhan 2020-12-01
I ran the `benchmark_incompressible_model.jl` script on the master branch (twice) and this branch (also twice), and 
actually see a tiny bit of a speedup, maybe only significant for larger CPU models though.

Hard to say whether it's noise, it might be more due to other processes causing small variations in runtime.

To me I don't think this PR slows down or speeds up the code, but it simplifies and improves the time stepping code so it should be merged.

There's a few more memory allocations now (due to extra kernel launches) but this shouldn't affect performance.

# System info

```
Oceananigans v0.44.1
Julia Version 1.5.2
Commit 539f3ce943 (2020-09-23 23:17 UTC)
Platform Info:
  OS: Linux (x86_64-pc-linux-gnu)
  CPU: Intel(R) Xeon(R) Silver 4214 CPU @ 2.20GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-9.0.1 (ORCJIT, cascadelake)
  GPU: TITAN V
```

# Master branch
```
                                        Incompressible model benchmarks
┌───────────────┬─────────────┬─────┬────────────┬────────────┬────────────┬────────────┬────────────┬────────┐
│ Architectures │ Float_types │  Ns │        min │     median │       mean │        max │     memory │ allocs │
├───────────────┼─────────────┼─────┼────────────┼────────────┼────────────┼────────────┼────────────┼────────┤
│           CPU │     Float32 │  32 │   5.399 ms │   5.668 ms │   5.758 ms │   7.186 ms │ 242.42 KiB │   1876 │
│           CPU │     Float32 │  64 │  36.710 ms │  37.583 ms │  37.974 ms │  41.678 ms │ 242.42 KiB │   1876 │
│           CPU │     Float32 │ 128 │ 312.780 ms │ 313.477 ms │ 313.622 ms │ 314.726 ms │ 242.42 KiB │   1876 │
│           CPU │     Float32 │ 256 │    2.802 s │    2.819 s │    2.819 s │    2.836 s │ 242.42 KiB │   1876 │
│           CPU │     Float64 │  32 │   5.828 ms │   6.049 ms │   6.157 ms │   7.044 ms │ 293.44 KiB │   1876 │
│           CPU │     Float64 │  64 │  43.084 ms │  43.619 ms │  43.650 ms │  44.363 ms │ 293.44 KiB │   1876 │
│           CPU │     Float64 │ 128 │ 365.051 ms │ 365.317 ms │ 365.475 ms │ 366.288 ms │ 293.44 KiB │   1876 │
│           CPU │     Float64 │ 256 │    3.602 s │    3.653 s │    3.653 s │    3.703 s │ 293.44 KiB │   1876 │
│           GPU │     Float32 │  32 │   2.797 ms │   2.870 ms │   2.918 ms │   3.435 ms │ 802.70 KiB │  11419 │
│           GPU │     Float32 │  64 │   3.120 ms │   3.207 ms │   3.300 ms │   4.224 ms │ 802.52 KiB │  11415 │
│           GPU │     Float32 │ 128 │   4.019 ms │   4.066 ms │   4.192 ms │   5.244 ms │ 802.52 KiB │  11415 │
│           GPU │     Float32 │ 256 │  15.942 ms │  23.497 ms │  22.763 ms │  23.588 ms │ 802.48 KiB │  11413 │
│           GPU │     Float64 │  32 │   3.079 ms │   3.166 ms │   3.226 ms │   3.728 ms │ 877.05 KiB │  11253 │
│           GPU │     Float64 │  64 │   3.458 ms │   3.522 ms │   3.591 ms │   3.981 ms │ 876.86 KiB │  11249 │
│           GPU │     Float64 │ 128 │   4.536 ms │   4.572 ms │   4.723 ms │   6.000 ms │ 876.58 KiB │  11231 │
│           GPU │     Float64 │ 256 │  21.794 ms │  32.107 ms │  31.073 ms │  32.198 ms │ 876.83 KiB │  11247 │
└───────────────┴─────────────┴─────┴────────────┴────────────┴────────────┴────────────┴────────────┴────────┘
```

```
      Incompressible model CPU -> GPU speedup
┌─────────────┬─────┬─────────┬─────────┬─────────┐
│ Float_types │  Ns │ speedup │  memory │  allocs │
├─────────────┼─────┼─────────┼─────────┼─────────┤
│     Float32 │  32 │  1.9749 │ 3.31118 │ 6.08689 │
│     Float32 │  64 │   11.72 │ 3.31041 │ 6.08475 │
│     Float32 │ 128 │ 77.0954 │ 3.31041 │ 6.08475 │
│     Float32 │ 256 │  119.98 │ 3.31028 │ 6.08369 │
│     Float64 │  32 │ 1.91043 │ 2.98887 │  5.9984 │
│     Float64 │  64 │ 12.3861 │ 2.98823 │ 5.99627 │
│     Float64 │ 128 │ 79.9049 │ 2.98727 │ 5.98667 │
│     Float64 │ 256 │ 113.772 │ 2.98813 │  5.9952 │
└─────────────┴─────┴─────────┴─────────┴─────────┘
```

# This branch

```
                                        Incompressible model benchmarks
┌───────────────┬─────────────┬─────┬────────────┬────────────┬────────────┬────────────┬────────────┬────────┐
│ Architectures │ Float_types │  Ns │        min │     median │       mean │        max │     memory │ allocs │
├───────────────┼─────────────┼─────┼────────────┼────────────┼────────────┼────────────┼────────────┼────────┤
│           CPU │     Float32 │  32 │   5.444 ms │   5.636 ms │   6.018 ms │   9.357 ms │ 247.69 KiB │   1916 │
│           CPU │     Float32 │  64 │  36.689 ms │  37.147 ms │  37.348 ms │  38.446 ms │ 247.69 KiB │   1916 │
│           CPU │     Float32 │ 128 │ 314.926 ms │ 316.673 ms │ 318.545 ms │ 338.621 ms │ 247.69 KiB │   1916 │
│           CPU │     Float32 │ 256 │    2.778 s │    2.781 s │    2.781 s │    2.783 s │ 247.69 KiB │   1916 │
│           CPU │     Float64 │  32 │   5.735 ms │   6.063 ms │   6.136 ms │   7.018 ms │ 299.80 KiB │   1916 │
│           CPU │     Float64 │  64 │  43.243 ms │  43.446 ms │  43.607 ms │  44.871 ms │ 299.80 KiB │   1916 │
│           CPU │     Float64 │ 128 │ 366.596 ms │ 367.479 ms │ 367.682 ms │ 369.125 ms │ 299.80 KiB │   1916 │
│           CPU │     Float64 │ 256 │    3.281 s │    3.331 s │    3.331 s │    3.381 s │ 299.80 KiB │   1916 │
│           GPU │     Float32 │  32 │   2.888 ms │   2.939 ms │   2.994 ms │   3.485 ms │ 814.47 KiB │  11744 │
│           GPU │     Float32 │  64 │   3.148 ms │   3.224 ms │   3.293 ms │   3.913 ms │ 814.28 KiB │  11740 │
│           GPU │     Float32 │ 128 │   4.002 ms │   4.089 ms │   4.210 ms │   5.287 ms │ 814.38 KiB │  11746 │
│           GPU │     Float32 │ 256 │  16.015 ms │  23.712 ms │  22.928 ms │  23.994 ms │ 814.38 KiB │  11746 │
│           GPU │     Float64 │  32 │   3.159 ms │   3.190 ms │   3.249 ms │   3.757 ms │ 892.39 KiB │  11578 │
│           GPU │     Float64 │  64 │   3.472 ms │   3.534 ms │   3.640 ms │   4.632 ms │ 892.20 KiB │  11574 │
│           GPU │     Float64 │ 128 │   4.479 ms │   4.537 ms │   4.700 ms │   6.206 ms │ 891.98 KiB │  11560 │
│           GPU │     Float64 │ 256 │  21.481 ms │  31.610 ms │  30.599 ms │  31.686 ms │ 892.30 KiB │  11580 │
└───────────────┴─────────────┴─────┴────────────┴────────────┴────────────┴────────────┴────────────┴────────┘
```

```
      Incompressible model CPU -> GPU speedup
┌─────────────┬─────┬─────────┬─────────┬─────────┐
│ Float_types │  Ns │ speedup │  memory │  allocs │
├─────────────┼─────┼─────────┼─────────┼─────────┤
│     Float32 │  32 │ 1.91728 │ 3.28829 │ 6.12944 │
│     Float32 │  64 │ 11.5212 │ 3.28753 │ 6.12735 │
│     Float32 │ 128 │ 77.4361 │ 3.28791 │ 6.13048 │
│     Float32 │ 256 │ 117.271 │ 3.28791 │ 6.13048 │
│     Float64 │  32 │ 1.90073 │ 2.97665 │  6.0428 │
│     Float64 │  64 │ 12.2925 │ 2.97603 │ 6.04071 │
│     Float64 │ 128 │    81.0 │  2.9753 │  6.0334 │
│     Float64 │ 256 │ 105.386 │ 2.97634 │ 6.04384 │
└─────────────┴─────┴─────────┴─────────┴─────────┘
```